(all): show currently selected value on next to sliders (come on google, give me a default handle I can use instead), use int instead of float

+++ app/src/genericweather/src/java/ui/activities/ConditionTargetConfigurationActivity.kt
@@ -64,7 +64,7 @@ class WeatherTargetConfigurationActivity : ComponentActivity() {
                             SmartspacerTargetProvider.notifyChange(context, GenericWeatherTarget::class.java)
                         },
                         range = (0..4),
-                        defaultPosition = dataPoints.toFloat()
+                        defaultPosition = dataPoints
                     )
                 }
             }
